+++ js/data.js
@@ -1,0 +1,35 @@
+const TITLES = [
+  'Домик у моря',
+  'Квартира с видом на залив',
+  'Уютная студия  центре города',
+  'Роскошный дворец',
+  'Комната в квартире с отличными соседями',
+];
+
+const TYPES = ['palace', 'flat', 'house', 'bungalow', 'hotel'];
+
+const CHECK_TIMES = ['12:00', '13:00', '14:00'];
+
+const DESCRIPTIONS = [
+  'Квартира в доме-памятнике купца Полежаева. В этом доме режиссер В. Бортко снимал в сцены для фильма по роману М.А. Булгакова «Мастер и Маргарита».',
+  'Квартира с прекрасным видом на Канал Грибоедова в доме начала XX века. В историческом центре, в непосредственной близости от главных достопримечательностей.',
+  'Старинный петербургский дом 1900 г. постройки. Совсем недавно в нём были реконструированы мини-студии, интересные исторические детали интерьера были бережно восстановлены.',
+];
+
+const FEATURES = [
+  'wifi',
+  'dishwasher',
+  'parking',
+  'washer',
+  'elevator',
+  'conditioner',
+];
+
+
+const PHOTOS = [
+  'https://assets.htmlacademy.ru/content/intensive/javascript-1/keksobooking/duonguyen-8LrGtIxxa4w.jpg',
+  'https://assets.htmlacademy.ru/content/intensive/javascript-1/keksobooking/brandon-hoogenboom-SNxQGWxZQi0.jpg',
+  'https://assets.htmlacademy.ru/content/intensive/javascript-1/keksobooking/claire-rendall-b6kAwr1i0Iw.jpg',
+];
+
+export {TITLES, TYPES, CHECK_TIMES, DESCRIPTIONS, FEATURES, PHOTOS};
